Refactor : 쿼리 로직 분리
